fix file type
